+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Added
 
-
+- Added GP Stroke From Spline node.
+- Added Spline From GP Stroke node.
 
 ### Fixed